DSP305x_01/Đồ án 1 image classification/serving/requirements.txt
fastapi
uvicorn
tensorflow-cpu
nest-asyncio
numpy
Pillow 
pybase64
python-multipart